Feature/office permissions

+++ src/TD.Komercijalno/TD.Komercijalno.Contracts/Dtos/Procedure/NabavnaCenaNaDanDto.cs
@@ -2,7 +2,7 @@
 {
     public class NabavnaCenaNaDanDto
     {
-        public int RobaId { get; set; }
+        public long RobaId { get; set; }
         public double NabavnaCenaBezPDV { get; set; }
     }
 }
