+++ src/TD.Komercijalno/TD.Komercijalno.Contracts/Entities/Komentar.cs
@@ -18,7 +18,7 @@ namespace TD.Komercijalno.Contracts.Entities
         public string? PrivatniKomentar { get; set; }
 
         [NotMapped]
-        public int Id { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+        public long Id { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         [NotMapped]
         public bool IsActive { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
         [NotMapped]
